fix: 리뷰 반영 — 5개 수정 (KAN-214)
- @AuthenticationPrincipal race condition: findByIdWithLock(PESSIMISTIC_WRITE)
- @CacheEvict → AFTER_COMMIT TransactionalEventListener 분리
- float → double (Account.companionScore + 연관 DTO/테스트)
- List<Object[]> → ScoreCountProjection named projection
- ErrorCode CR_003 메시지 수정 ("참여자가 아니면 리뷰를 작성할 수 없습니다")
- Flyway V202606041431: companion_score FLOAT → DOUBLE ALTER

+++ src/main/java/com/chunbaetour/domain/auth/Account.java
@@ -49,7 +49,7 @@ public class Account {
     private String language;
 
     @Column(name = "companion_score", nullable = false)
-    private float companionScore;
+    private double companionScore;
 
     @Column(name = "companion_review_count", nullable = false)
     private int companionReviewCount;
@@ -270,7 +270,7 @@ public class Account {
      */
     // 동행 리뷰 등록 시 증분식 평균 갱신 — (currentScore * count + newScore) / (count + 1)
     public void addCompanionReview(int score) {
-        this.companionScore = (this.companionScore * this.companionReviewCount + score) / (this.companionReviewCount + 1);
+        this.companionScore = (this.companionScore * this.companionReviewCount + score) / (double) (this.companionReviewCount + 1);
         this.companionReviewCount++;
     }
